New file_dialog file for dialogs
