+++ src/Dosepy/tools/controller.py
@@ -1,17 +1,14 @@
-"""Class used as a controller in a MVC pattern."""
+"""Class used as a controller in a ModelViewControll (MVC) pattern."""
 
 from PySide6.QtWidgets import (
     QFileDialog,
-    QDialog,
-    QDialogButtonBox,
-    QLabel,
-    QVBoxLayout,
     QHeaderView,
 )
-from PySide6.QtCore import QDir
 from pathlib import Path
 import os
 import numpy as np
+
+from gui_widgets.file_dialog import open_files_dialog, Error_Dialog
 
 class DosepyController():
     def __init__(self, model, view):
@@ -25,54 +22,45 @@ class DosepyController():
     # Related to Calibration
         
     def _open_file_button(self):
-        #print("Hola boton open")
-        dialog = QFileDialog()
-        #dialog.setDirectory(r'C:\images')
-        dialog.setDirectory(QDir.home())
-        dialog.setFileMode(QFileDialog.FileMode.ExistingFiles)
-        dialog.setNameFilter("Images (*.tif *.tiff)")
+        new_files = open_files_dialog("Images (*.tif *.tiff)")
 
-        if dialog.exec():
-            filenames = dialog.selectedFiles()
-            # Absolute paths to files
-            list_files = \
-                [str(Path(filename)) for filename in filenames] + \
-                self._view.cal_widget.get_files_list()
-            
-            if filenames:
-                if self._model.are_valid_tif_files(list_files):
-                    if self._model.are_files_equal_shape(list_files):
+        if new_files:
+    
+            if self._model.are_valid_tif_files(new_files):
+                current_files = self._view.cal_widget.get_files_list()
+                list_files = current_files + new_files
+                if self._model.are_files_equal_shape(list_files):
 
-                        # Display path to files
-                        self._view.cal_widget.set_files_list(list_files)
+                    # Display path to files
+                    self._view.cal_widget.set_files_list(list_files)
 
-                        # load the files
-                        self._model.calibration_img = self._model.load_files(
-                            list_files,
-                            for_calib=True
-                            )
-                        self._view.cal_widget.plot_image(self._model.calibration_img)
+                    # load the files
+                    self._model.calibration_img = self._model.load_files(
+                        list_files,
+                        for_calib=True
+                        )
+                    self._view.cal_widget.plot_image(self._model.calibration_img)
 
-                        # Find how many film we have and show a table for user input dose values
-                        self._model.calibration_img.set_labeled_img()
-                        num = self._model.calibration_img.number_of_films
-                        print(f"Number of detected films: {num}")
-                        self._view.cal_widget.set_table_rows(rows = num)
-                        header = self._view.cal_widget.dose_table.horizontalHeader()
-                        self._view.cal_widget.dose_table.cellChanged.connect(self._is_a_valid_dose)
-                        header.setSectionResizeMode(0, QHeaderView.ResizeMode.Stretch)
+                    # Find how many film we have and show a table for user input dose values
+                    self._model.calibration_img.set_labeled_img()
+                    num = self._model.calibration_img.number_of_films
+                    print(f"Number of detected films: {num}")
+                    self._view.cal_widget.set_table_rows(rows = num)
+                    header = self._view.cal_widget.dose_table.horizontalHeader()
+                    self._view.cal_widget.dose_table.cellChanged.connect(self._is_a_valid_dose)
+                    header.setSectionResizeMode(0, QHeaderView.ResizeMode.Stretch)
 
-                        self._view.cal_widget.apply_button.setEnabled(True)
-                    
-                    else:
-                        msg = "The tiff files must have the same shape."
-                        print(msg)
-                        Error_Dialog(msg).exec()
-
+                    self._view.cal_widget.apply_button.setEnabled(True)
+                
                 else:
-                    msg = "Invalid file. Is it a tiff RGB file?"
+                    msg = "The tiff files must have the same shape."
                     print(msg)
                     Error_Dialog(msg).exec()
+
+            else:
+                msg = "Invalid file. Is it a tiff RGB file?"
+                print(msg)
+                Error_Dialog(msg).exec()
 
     def _apply_calib_button(self):
         print("Apply button pressed.")
@@ -101,7 +89,7 @@ class DosepyController():
                 print(msg)
                 Error_Dialog(msg).exec()
         else:
-            msg = "Invalid dose values"
+            msg = "Invalid dose values."
             print(msg)
             Error_Dialog(msg).exec()
 
@@ -152,29 +140,6 @@ class DosepyController():
         self._view.cal_widget.apply_button.clicked.connect(self._apply_calib_button)
         self._view.cal_widget.save_cal_button.clicked.connect(self._save_calib_button)
         #self._view.cal_widget.clear_button.clicked.connect(self._clear_file_button) #TODO_
+
+        #self._view.
             
-
-class Error_Dialog(QDialog):
-    """
-    Basic QDialog to show an error message.
-    
-    Parameters
-    ----------
-    msg : str
-        The message to show.
-    """
-    def __init__(self, msg):
-        super().__init__()
-
-        self.setWindowTitle("Error")
-
-        buttons = (QDialogButtonBox.StandardButton.Ok)
-        
-        self.buttonBox = QDialogButtonBox(buttons)
-        self.buttonBox.accepted.connect(self.accept)
-
-        self.layout = QVBoxLayout()
-        message = QLabel(msg)
-        self.layout.addWidget(message)
-        self.layout.addWidget(self.buttonBox)
-        self.setLayout(self.layout)
